flask_backend/requirements.txt
flask==3.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
openai>=1.3.0
httpx>=0.28.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
redis==5.0.1
pydantic>=2.8.0
langgraph>=0.2.0
langchain>=0.3.27
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.2.0 
boto3>=1.40.9
